+++ src/main/java/kakaopay/housingfinance/controller/HousingFinanceController.java
@@ -1,10 +1,12 @@
 package kakaopay.housingfinance.controller;
 
+import kakaopay.housingfinance.service.BankService;
 import kakaopay.housingfinance.service.FinanceService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -13,13 +15,20 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.Map;
 
 @RestController
-public class FinanceController {
-    private static Logger logger = LoggerFactory.getLogger(FinanceController.class);
+public class HousingFinanceController {
+    private static Logger logger = LoggerFactory.getLogger(HousingFinanceController.class);
     private static final HttpHeaders httpHeaders = new HttpHeaders();
     private final FinanceService financeService;
+    private final BankService bankService;
 
-    public FinanceController(FinanceService financeService) {
+    public HousingFinanceController(FinanceService financeService, BankService bankService) {
+        httpHeaders.setContentType(MediaType.APPLICATION_JSON_UTF8);
         this.financeService = financeService;
+        this.bankService = bankService;
+    }
+    @GetMapping("/banks")
+    public ResponseEntity<Map<String,Object>> getAllBanks(){
+        return new ResponseEntity<>(bankService.getAllBanks(),httpHeaders, HttpStatus.OK);
     }
 
     @GetMapping("/finance/status")
